move HapiPlugin, clean tests
