+++ src/plugin/hapi.js
@@ -1,6 +1,6 @@
-import { processLogData } from '../helpers';
 import Syslog from '../adapter/syslog';
 import Console from '../adapter/console';
+import HapiPlugin from './hapi-plugin.js';
 
 /**
  *
@@ -69,83 +69,4 @@ export function setupLogger(Logger, server, options) {
         type: 'onPostStart',
         method: (server, next) => hapiPlugin.onServerStart(server, next)
     });
-}
-
-/**
- * HapiPlugin
- *
- */
-export class HapiPlugin {
-
-    /**
-     *
-     * @param {Logger} Logger
-     * @constructor
-     */
-    constructor(Logger) {
-        this._logger = Logger;
-    }
-
-    /**
-     *
-     * @param {Server} server
-     * @param next
-     */
-    async onServerStart(server, next) {
-        this._logger.info('server started:', server.info.uri);
-
-        if (server.version < '17.0.0') {
-            return next();
-        }
-    }
-
-    /**
-     *
-     * @param {Server} server
-     * @param {Object} event
-     * @param {Array} tags
-     */
-    onServerLog(server, event, tags) {
-        const data = event.error || event.data;
-        const { severity, message } = processLogData(data, tags);
-        this._logger[severity](message);
-    }
-
-    /**
-     *
-     * @param {Request} request
-     * @param {Object} event
-     * @param {Array} tags
-     */
-    onRequestLog(request, event, tags) {
-        const data = event.error || event.data;
-        const { severity, message } = processLogData(data, tags);
-        this._logger[severity](message);
-    }
-
-    /**
-     *
-     * @param {Request} request
-     * @param {error} err
-     */
-    onRequestError(request, err) {
-        this._logger.error(err);
-    }
-
-    /**
-     *
-     * @param {Request} request
-     * @param {Object} event
-     * @param {Object} tags
-     */
-    onRequestInternal(request, event, tags) {
-        // new request just received
-        if (tags.received) {
-            const method = request.method.toUpperCase();
-            const proto = (request.headers['x-forwarded-proto'] || request.connection.info.protocol);
-            const url = `${proto}://${request.info.host}${request.url.path}`;
-            this._logger.debug(method, url);
-        }
-    }
-
 }
